+++ examples/tests/test_macos.py
@@ -1,0 +1,183 @@
+import sys
+import time
+import uuid
+
+import pytest
+from media import Recvonly, Sendonly
+
+"""
+GitHub Actions で Video Toolbox を送受信で利用しようとするとエラーになるので、
+テストを sendonly のみに絞っている
+"""
+
+
+def test_macos_h264_sendonly(setup):
+    signaling_urls = setup.get("signaling_urls")
+    channel_id_prefix = setup.get("channel_id_prefix")
+    metadata = setup.get("metadata")
+
+    channel_id = f"{channel_id_prefix}_{__name__}_{sys._getframe().f_code.co_name}_{uuid.uuid4()}"
+
+    sendonly = Sendonly(
+        signaling_urls,
+        channel_id,
+        metadata=metadata,
+        video_codec_type="H264",
+    )
+    sendonly.connect()
+
+    time.sleep(5)
+
+    sendonly_stats = sendonly.get_stats()
+
+    # codec が無かったら StopIteration 例外が上がる
+    codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
+    # H.264 が採用されているかどうか確認する
+    assert codec_stats["mimeType"] == "video/H264"
+
+    # outbound-rtp が無かったら StopIteration 例外が上がる
+    outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
+    assert outbound_rtp_stats["encoderImplementation"] == "VideoToolbox"
+    assert outbound_rtp_stats["bytesSent"] > 0
+    assert outbound_rtp_stats["packetsSent"] > 0
+
+    sendonly.disconnect()
+
+
+def test_macos_h265_sendonly(setup):
+    signaling_urls = setup.get("signaling_urls")
+    channel_id_prefix = setup.get("channel_id_prefix")
+    metadata = setup.get("metadata")
+
+    channel_id = f"{channel_id_prefix}_{__name__}_{sys._getframe().f_code.co_name}_{uuid.uuid4()}"
+
+    sendonly = Sendonly(
+        signaling_urls,
+        channel_id,
+        metadata,
+        video_codec_type="H265",
+    )
+    sendonly.connect()
+
+    time.sleep(5)
+
+    sendonly_stats = sendonly.get_stats()
+
+    # codec が無かったら StopIteration 例外が上がる
+    codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
+    # H.264 が採用されているかどうか確認する
+    assert codec_stats["mimeType"] == "video/H265"
+
+    # outbound-rtp が無かったら StopIteration 例外が上がる
+    outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
+    assert outbound_rtp_stats["encoderImplementation"] == "VideoToolbox"
+    assert outbound_rtp_stats["bytesSent"] > 0
+    assert outbound_rtp_stats["packetsSent"] > 0
+
+    sendonly.disconnect()
+
+
+@pytest.mark.skip(reason="ローカルでは成功する")
+def test_macos_h264_sendonly_recvonly(setup):
+    signaling_urls = setup.get("signaling_urls")
+    channel_id_prefix = setup.get("channel_id_prefix")
+    metadata = setup.get("metadata")
+
+    channel_id = f"{channel_id_prefix}_{__name__}_{sys._getframe().f_code.co_name}_{uuid.uuid4()}"
+
+    sendonly = Sendonly(
+        signaling_urls,
+        channel_id,
+        metadata,
+        video_codec_type="H264",
+    )
+    sendonly.connect()
+
+    recvonly = Recvonly(
+        signaling_urls,
+        channel_id,
+        metadata=metadata,
+    )
+    recvonly.connect()
+
+    time.sleep(5)
+
+    sendonly_stats = sendonly.get_stats()
+    recvonly_stats = recvonly.get_stats()
+
+    # codec が無かったら StopIteration 例外が上がる
+    sendonly_codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
+    # H.264 が採用されているかどうか確認する
+    assert sendonly_codec_stats["mimeType"] == "video/H264"
+
+    # outbound-rtp が無かったら StopIteration 例外が上がる
+    outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
+    assert outbound_rtp_stats["encoderImplementation"] == "VideoToolbox"
+    assert outbound_rtp_stats["bytesSent"] > 0
+    assert outbound_rtp_stats["packetsSent"] > 0
+
+    # codec が無かったら StopIteration 例外が上がる
+    recvonly_codec_stats = next(s for s in recvonly_stats if s.get("type") == "codec")
+    # H.264 が採用されているかどうか確認する
+    assert recvonly_codec_stats["mimeType"] == "video/H264"
+
+    # outbound-rtp が無かったら StopIteration 例外が上がる
+    inbound_rtp_stats = next(s for s in recvonly_stats if s.get("type") == "inbound-rtp")
+    assert inbound_rtp_stats["decoderImplementation"] == "VideoToolbox"
+    assert inbound_rtp_stats["bytesReceived"] > 0
+    assert inbound_rtp_stats["packetsReceived"] > 0
+
+    sendonly.disconnect()
+    recvonly.disconnect()
+
+
+@pytest.mark.skip(reason="ローカルでは成功する")
+def test_macos_h265_sendonly_recvonly(setup):
+    signaling_urls = setup.get("signaling_urls")
+    channel_id_prefix = setup.get("channel_id_prefix")
+    metadata = setup.get("metadata")
+
+    channel_id = f"{channel_id_prefix}_{__name__}_{sys._getframe().f_code.co_name}_{uuid.uuid4()}"
+
+    sendonly = Sendonly(
+        signaling_urls,
+        channel_id,
+        metadata,
+        video_codec_type="H265",
+    )
+    sendonly.connect()
+
+    recvonly = Recvonly(
+        signaling_urls,
+        channel_id,
+        metadata=metadata,
+    )
+    recvonly.connect()
+
+    time.sleep(5)
+
+    sendonly_stats = sendonly.get_stats()
+    recvonly_stats = recvonly.get_stats()
+
+    # codec が無かったら StopIteration 例外が上がる
+    sendonly_codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
+    assert sendonly_codec_stats["mimeType"] == "video/H265"
+
+    # outbound-rtp が無かったら StopIteration 例外が上がる
+    outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
+    assert outbound_rtp_stats["encoderImplementation"] == "VideoToolbox"
+    assert outbound_rtp_stats["bytesSent"] > 0
+    assert outbound_rtp_stats["packetsSent"] > 0
+
+    # codec が無かったら StopIteration 例外が上がる
+    recvonly_codec_stats = next(s for s in recvonly_stats if s.get("type") == "codec")
+    assert recvonly_codec_stats["mimeType"] == "video/H265"
+
+    # outbound-rtp が無かったら StopIteration 例外が上がる
+    inbound_rtp_stats = next(s for s in recvonly_stats if s.get("type") == "inbound-rtp")
+    assert inbound_rtp_stats["decoderImplementation"] == "VideoToolbox"
+    assert inbound_rtp_stats["bytesReceived"] > 0
+    assert inbound_rtp_stats["packetsReceived"] > 0
+
+    sendonly.disconnect()
+    recvonly.disconnect()
